requirements.txt
Flask
supabase
python-dotenv
colorama
Werkzeug
click
itsdangerous
Jinja2
MarkupSafe